Replace DisplayInfo-based API with qualifier strings.
